fix: change processor_name from "numeric" to "integer" to be consistent with the docstring.

+++ src/calt/data_loader/data_loader.py
@@ -82,7 +82,7 @@ def load_data(
             max_degree=max_degree,
             max_coeff=max_coeff,
         )
-    elif processor_name == "numeric":
+    elif processor_name == "integer":
         preprocessor = IntegerToInternalProcessor(max_coeff=max_coeff)
     else:
         raise ValueError(f"Unknown processor: {processor_name}")
